refactor: remove 50 dead functions unreachable from main
Whole-program reachability analysis (golang.org/x/tools/cmd/deadcode)
found 50 functions unreachable from main that the unused linter cannot
flag: it assumes exported identifiers have external importers (this is
a CLI; they do not) and counts test references as usage (18 of the 50
were kept alive only by their own tests).

Removed across gitlab (26), cmd (8), configuration (8), and
github/control/internal/provider (8), along with orphaned tests,
helpers, and imports. Tests that depended on dead functions were
rewired onto live paths instead of deleted:
- policies tests now call ScanGitHubWorkflowsWithProgress, the live
  collector, instead of its stale diverged duplicate ScanGitHubWorkflows
- Engine.LoadFromFS callers use LoadFromFSFiltered(fs, nil)
- cmd wizard drift-guard tests rebuilt against the live
  initWizardState.toPlumberConfig() path
- SplitImageRef coverage moved to the live unexported splitImageRef;
  the alias-normalization E2E now exercises the real GitHub scan path,
  pins the extracted folded image, and adds an untrusted-image positive
  control so the check cannot pass vacuously
