+++ configuration/expression_test.go
@@ -213,133 +213,3 @@ func TestParseRequiredExpression(t *testing.T) {
 	}
 }
 
-func TestGroupsToExpression(t *testing.T) {
-	tests := []struct {
-		name     string
-		groups   [][]string
-		expected string
-	}{
-		{
-			name:     "empty",
-			groups:   [][]string{},
-			expected: "",
-		},
-		{
-			name:     "nil",
-			groups:   nil,
-			expected: "",
-		},
-		{
-			name:     "single item single group",
-			groups:   [][]string{{"a"}},
-			expected: "a",
-		},
-		{
-			name:     "multiple items single group",
-			groups:   [][]string{{"a", "b", "c"}},
-			expected: "a AND b AND c",
-		},
-		{
-			name:     "single items multiple groups",
-			groups:   [][]string{{"a"}, {"b"}, {"c"}},
-			expected: "a OR b OR c",
-		},
-		{
-			name:     "mixed groups",
-			groups:   [][]string{{"a", "b"}, {"c"}},
-			expected: "(a AND b) OR c",
-		},
-		{
-			name:     "two multi-item groups",
-			groups:   [][]string{{"a", "b"}, {"c", "d"}},
-			expected: "(a AND b) OR (c AND d)",
-		},
-		{
-			name:     "skip empty groups",
-			groups:   [][]string{{}, {"a"}, {}},
-			expected: "a",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got := GroupsToExpression(tt.groups)
-			if got != tt.expected {
-				t.Errorf("GroupsToExpression(%v)\n  got:  %q\n  want: %q", tt.groups, got, tt.expected)
-			}
-		})
-	}
-}
-
-// TestRoundTrip verifies that parsing an expression and converting back gives
-// an equivalent (though not necessarily identical) expression.
-func TestRoundTrip(t *testing.T) {
-	tests := []struct {
-		name string
-		expr string
-	}{
-		{"single", "a"},
-		{"two AND", "a AND b"},
-		{"two OR", "a OR b"},
-		{"DNF", "(a AND b) OR c"},
-		{"complex DNF", "(a AND b) OR (c AND d)"},
-		{"real world", "(components/sast/sast AND components/secret-detection/secret-detection) OR your-org/full-security/full-security"},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			// Parse expression to groups
-			groups, err := ParseRequiredExpression(tt.expr)
-			if err != nil {
-				t.Fatalf("ParseRequiredExpression(%q) failed: %v", tt.expr, err)
-			}
-
-			// Convert groups back to expression
-			regenerated := GroupsToExpression(groups)
-
-			// Parse the regenerated expression
-			groups2, err := ParseRequiredExpression(regenerated)
-			if err != nil {
-				t.Fatalf("ParseRequiredExpression(%q) [regenerated] failed: %v", regenerated, err)
-			}
-
-			// The DNF groups should be identical
-			if !reflect.DeepEqual(groups, groups2) {
-				t.Errorf("Round-trip failed for %q:\n  groups1: %v\n  regen:   %q\n  groups2: %v",
-					tt.expr, groups, regenerated, groups2)
-			}
-		})
-	}
-}
-
-func TestValidateExpression(t *testing.T) {
-	valid := []string{
-		"",
-		"a",
-		"a AND b",
-		"a OR b",
-		"(a AND b) OR c",
-		"a AND (b OR c)",
-	}
-	for _, expr := range valid {
-		if err := ValidateExpression(expr); err != nil {
-			t.Errorf("ValidateExpression(%q) expected valid, got error: %v", expr, err)
-		}
-	}
-
-	invalid := []string{
-		"AND",
-		"OR",
-		"a AND",
-		"a OR",
-		"(a",
-		"a)",
-		"()",
-		"a AND AND b",
-	}
-	for _, expr := range invalid {
-		if err := ValidateExpression(expr); err == nil {
-			t.Errorf("ValidateExpression(%q) expected error, got nil", expr)
-		}
-	}
-}
