+++ github/github_metadata.go
@@ -129,25 +129,18 @@ type repoCacheEntry struct {
 	err      error
 }
 
-// NewGitHubMetadataClient builds a client using the gh-CLI auth
+// NewGitHubMetadataClientForHost builds a client using the gh-CLI auth
 // store. Returns a usable client even when authentication is
 // missing — see Available() to check. Honors the
 // PLUMBER_DISABLE_GITHUB_API env var which short-circuits the
 // client into degraded mode regardless of auth state.
 //
-// Targets api.github.com by default. For GitHub Enterprise Server
-// instances, use NewGitHubMetadataClientForHost with the GHES API
-// host (e.g. "ghes.example.com" or "ghes.example.com/api/v3").
-func NewGitHubMetadataClient() *GitHubMetadataClient {
-	return NewGitHubMetadataClientForHost("")
-}
-
-// NewGitHubMetadataClientForHost is the GHES-aware constructor. When
-// host is empty the client targets api.github.com via the default
+// When host is empty the client targets api.github.com via the default
 // go-gh resolution chain (gh auth, GH_TOKEN, GITHUB_TOKEN). When
-// host is non-empty the client is bound to that host — pair with a
-// GH_TOKEN (or GH_ENTERPRISE_TOKEN) that has access to the GHES
-// instance.
+// host is non-empty the client is bound to that host — for GitHub
+// Enterprise Server pass the GHES API host (e.g. "ghes.example.com"
+// or "ghes.example.com/api/v3") paired with a GH_TOKEN (or
+// GH_ENTERPRISE_TOKEN) that has access to the GHES instance.
 func NewGitHubMetadataClientForHost(host string) *GitHubMetadataClient {
 	c := &GitHubMetadataClient{
 		apiBaseURL:    apiBaseURLForHost(host),
