+++ gitlab/request.go
@@ -410,8 +410,8 @@ func GetGitlabInstanceVariables(token string, instanceUrl string, conf *configur
 
 // GetGitlabCIComponentResource fetches a SINGLE CI/CD catalog resource by its
 // project full path, with its released versions and their components. It is the
-// targeted replacement for the instance-wide GetGitlabCIComponentResources
-// enumeration: on a large catalog (gitlab.com has ~800 resources) that broad
+// targeted replacement for the removed instance-wide catalog enumeration: on
+// a large catalog (gitlab.com has ~800 resources) that broad
 // query returns a multi-megabyte payload that exceeds the HTTP timeout, gets
 // cancelled, and silently disables outdated-include detection (#156). This
 // per-component lookup returns in well under a second and is complete (no
@@ -495,118 +495,4 @@ func GetGitlabCIComponentResource(fullPath string, token string, instanceUrl str
 		res.Versions = append(res.Versions, version)
 	}
 	return res, nil
-}
-
-// GetGitlabCIComponentResources fetches all CI component resources from GitLab.
-//
-// Deprecated: this instance-wide enumeration is too heavy on large catalogs and
-// is no longer used for version resolution (see GetGitlabCIComponentResource
-// and #156). Kept for any external callers.
-func GetGitlabCIComponentResources(isGroup bool, token string, instanceUrl string, conf *configuration.Configuration) ([]CICatalogResource, error) {
-	l := logrus.WithFields(logrus.Fields{
-		"action":      "GetGitlabCIComponentResources",
-		"instanceUrl": instanceUrl,
-	})
-
-	scope := "ALL"
-	if isGroup {
-		scope = "NAMESPACES"
-	}
-
-	query := fmt.Sprintf(`
-	query getCIComponentResources {
-		ciCatalogResources(scope: %s){
-			nodes {
-				id
-				name
-				fullPath
-				webPath
-				versions{
-					nodes{
-						name
-						components {
-							nodes {
-								id
-								name
-								includePath
-							}
-						}
-					}
-				}
-			}
-		}
-	}`, scope)
-
-	graphqlClient := GetGraphQLClient(instanceUrl, conf)
-	req := graphql.NewRequest(query)
-	req.Header.Set("Authorization", fmt.Sprintf("Bearer %s", token))
-
-	type componentNode struct {
-		ID          string `json:"id"`
-		Name        string `json:"name"`
-		IncludePath string `json:"includePath"`
-	}
-
-	type componentsNodes struct {
-		Nodes []componentNode `json:"nodes"`
-	}
-
-	type versionNode struct {
-		Name       string          `json:"name"`
-		Path       string          `json:"path"`
-		Components componentsNodes `json:"components"`
-	}
-
-	type versionsNodes struct {
-		Nodes []versionNode `json:"nodes"`
-	}
-
-	type resourceNode struct {
-		ID       string        `json:"id"`
-		Name     string        `json:"name"`
-		FullPath string        `json:"fullPath"`
-		WebPath  string        `json:"webPath"`
-		Versions versionsNodes `json:"versions"`
-	}
-
-	type ciResourcesResponse struct {
-		CICatalogResources struct {
-			Nodes []resourceNode `json:"nodes"`
-		} `json:"ciCatalogResources"`
-	}
-
-	var graphqlResp ciResourcesResponse
-	if err := graphqlClient.Run(context.Background(), req, &graphqlResp); err != nil {
-		l.WithError(err).Error("Failed to execute GraphQL query")
-		return nil, err
-	}
-
-	resources := make([]CICatalogResource, 0, len(graphqlResp.CICatalogResources.Nodes))
-	for _, node := range graphqlResp.CICatalogResources.Nodes {
-		resource := CICatalogResource{
-			ID:       node.ID,
-			Name:     node.Name,
-			FullPath: node.FullPath,
-			WebPath:  node.WebPath,
-			Versions: make([]CICatalogResourceVersion, 0, len(node.Versions.Nodes)),
-		}
-
-		for _, vNode := range node.Versions.Nodes {
-			version := CICatalogResourceVersion{
-				Name:       vNode.Name,
-				Path:       vNode.Path,
-				Components: make([]CIComponent, 0, len(vNode.Components.Nodes)),
-			}
-
-			for _, cNode := range vNode.Components.Nodes {
-				version.Components = append(version.Components, CIComponent(cNode))
-			}
-
-			resource.Versions = append(resource.Versions, version)
-		}
-
-		resources = append(resources, resource)
-	}
-
-	return resources, nil
 }
